requirements.txt
chardet
utils
python-dateutil==2.8.0
django
pandas
plotly
psutil
requests